relocate admin panel UI files
