--- admin-panel/src/app/layout.js
@@ -1,17 +1,0 @@
-import { Inter } from 'next/font/google';
-import './globals.css';
-
-const inter = Inter({ subsets: ['latin'] });
-
-export const metadata = {
-  title: 'Admin Panel',
-  description: 'admin panel for managing your job listings',
-};
-
-export default function RootLayout({ children }) {
-  return (
-    <html lang="en">
-      <body className={inter.className}>{children}</body>
-    </html>
-  );
-}
