+++ src/app/admin-panel/home/layout.js
@@ -1,5 +1,5 @@
 import { Inter } from 'next/font/google';
-import '../globals.css';
+import '../style.css';
 
 const inter = Inter({
   subsets: ['latin'],
