+++ src/components/ui/card.jsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/libs/utils';
 
 const Card = React.forwardRef(({ className, ...props }, ref) => (
   <div
